+++ src/QuadcopterColor.hpp
@@ -12,7 +12,7 @@ private:
 	
 public:
 	QuadcopterColor(cv::Scalar minColor, cv::Scalar maxColor, int id);
-	QuadcopterColor(int minHue, int maxHue, int minSaturation, int maxSaturation, int minValue, int maxValue, int id);
+	QuadcopterColor(uint8_t minHue, uint8_t maxHue, uint8_t minSaturation, uint8_t maxSaturation, uint8_t minValue, uint8_t maxValue, int id);
 	cv::Scalar getMinColor();
 	cv::Scalar getMaxColor();
 	int getId();
